check message types for ConI input proto parameters (#129)
Calling methods like data-update should check the input parameter Message type.
When those methods are called it may happen that accidentialy it is called with a ods.DataMatrix instead of ods.DataMtrices. This is very hard to be determined/debugged. Its would be easy to check the parameter for correct type.