set allow_redirects to False to all requests calls by default (#125)
the python requests library enables redirect by default.
This means if a server answers with 302 or 308 it will follow the given url.
Because this will happen on each call it is no good idea to swith this on.

e.g. Server is returning http as con_i instead of https and redirect is switched on.